Resync
-flare color now synced with other clients
-updated ping textures
-updated/added flare textures
-fixed fuel refinery voxelshape
-rewrote zipline item code
-you can now change the zipline direction while moving
-fixed zipline suffocate in block bug
-added satchel charge detonator item
-added missing translations
-added zipline string rendering
-added fuel tooltip
-fixed ommoran hearthstone locator select bug
-fixed ingredients gen... again
-improved liquid morkite lake gen
-optimized dwarf class reading system
-added misc tools can only be used by dwarfs (not leaf lovers)
-fixed zipline server crash
-fixed zipline string entity not removed after block removing
-fixed flare gun disperses after shoot
-fixed grappling hook hook not kill bug
-tweaked fuel refill item
-removed fuel canister
-improved fuel canister fuel system
-updated fuel canister recipe
-fixed drills not accept fuel
-fuel refinery, class changer must now be on sold block
-fixed platform gun shoot prevented some block placing
-added fuel refinery refuel mechanics (refill refinery, refill fuel)
-fixed fuel refinery crash
-set liquid morkite stack size to 1
-fixed blockentity data still exists after block destroy
-fixed zipline offset
-added tooltip to drills
-fixed zipline movement
-fixed finally zipline string rendering
-fixed tool use animation
-added salute requires to be a dwarf message
-fixed ommoran hearthstone structure model
-fixed ommoran hearthstone generation
-player now see his own class activated on the class changer
-changed lootbugs can now only be hurt when pacifist mode is false
-improved config
-added config to disable satchel charge blink anim (blink COULD produce lag when too many satchel charges are placed)
-fixed lootbug pacifist mode
-fixed explosive diarrhoea crash
-improved freeze effect
-removed dafuck item (it does not match with lootdebugs)
-rearranged class changer buttons
-removed player inventory from class changer
-fixed lootbug gold blockdrop
-fixed class changer leave ghost block behind after removed block under it
-fixed MASSIVE Development Engine core crash
-satchelcharge detonator now delete itself when no connection to satchel charge
-fixed class null crash
-removed rickroll disc because, why?
-fixed ping offset
-added new ping sounds
-sorted translation file
-added ping effect sound
-updated plascrete foam, flare sound
-added new block sounds
-fixed block sounds
-removed dafuck recipe
-fixed creeper drop music disc tag
-added drunkness movement
-rewrote fuel refinery fluid code
-added fuel refinery gui
-rewrote fuel refinery fuel calculating
-rewrote packet message system (upgraded to newer system)
-changed zipline end hook
-fixed zipline bugs

+++ src/main/java/net/the_goldbeards/lootdebugs/Items/Tools/CompanyStandardPickaxe/CompanyPickAxeItem.java
@@ -1,11 +1,15 @@
 package net.the_goldbeards.lootdebugs.Items.Tools.CompanyStandardPickaxe;
 
+import net.minecraft.ChatFormatting;
 import net.minecraft.core.BlockPos;
 import net.minecraft.core.particles.ParticleTypes;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.sounds.SoundSource;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResult;
 import net.minecraft.world.damagesource.DamageSource;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.ItemStack;
@@ -16,7 +20,10 @@ import net.minecraft.world.item.enchantment.Enchantment;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.minecraft.world.level.block.Blocks;
-import net.the_goldbeards.lootdebugs.Sound.ModSounds;
+import net.minecraft.world.level.block.state.BlockState;
+import net.the_goldbeards.lootdebugs.init.Sound.ModSounds;
+import net.the_goldbeards.lootdebugs.capability.Class.IClassData;
+import net.the_goldbeards.lootdebugs.util.UsefullStuff;
 
 public class CompanyPickAxeItem extends PickaxeItem  {
 
@@ -35,16 +42,15 @@ public class CompanyPickAxeItem extends PickaxeItem  {
     public InteractionResult interactLivingEntity(ItemStack pStack, Player pPlayer, LivingEntity pInteractionTarget, InteractionHand pUsedHand) {
 
 
-        if(!pPlayer.getCooldowns().isOnCooldown(this))
+        if(!pPlayer.getCooldowns().isOnCooldown(this) && !UsefullStuff.DwarfClasses.canPlayerUseItem(pStack, pPlayer, IClassData.Classes.LeafLover))
         {
             BlockPos pos = pInteractionTarget.getOnPos();
-            pInteractionTarget.hurt(DamageSource.GENERIC, 10);
-            pPlayer.level.playSound(pPlayer,pos, ModSounds.POWERATTACK.get(), SoundSource.PLAYERS, 35,1);
+            pInteractionTarget.hurt(DamageSource.GENERIC, 20);
+            pPlayer.level.playSound(null, pos, ModSounds.POWERATTACK.get(), SoundSource.PLAYERS, 35,1);
             pPlayer.level.addParticle(ParticleTypes.CRIT, pos.getX(), pos.getY(),pos.getZ(),1,1,1);
             pPlayer.getCooldowns().addCooldown(this, 800);
             return InteractionResult.SUCCESS;
         }
-
 
 
         return InteractionResult.PASS;
@@ -63,32 +69,48 @@ public class CompanyPickAxeItem extends PickaxeItem  {
         CompanyPickAxeItem item = (CompanyPickAxeItem) pContext.getItemInHand().getItem();
 
 
+        if(!UsefullStuff.DwarfClasses.canPlayerUseItem(itemstack, player, IClassData.Classes.LeafLover))
+        {
 
-                if (block != Blocks.BEDROCK || block != Blocks.COMMAND_BLOCK) {
-                    player.getCooldowns().addCooldown(this,800);
-                    level.playSound(player,pos, ModSounds.POWERATTACK.get(), SoundSource.PLAYERS, 35,1);
-                    level.addParticle(ParticleTypes.CRIT, pos.getX(), pos.getY(),pos.getZ(),1,1,1);
-                    level.destroyBlock(pos, true);
-                    level.updateNeighborsAt(pos, level.getBlockState(pos).getBlock());
+            if (block != Blocks.BEDROCK || block != Blocks.COMMAND_BLOCK) {
+                player.getCooldowns().addCooldown(this,800);
+                level.playSound(player,pos, ModSounds.POWERATTACK.get(), SoundSource.PLAYERS, 35,1);
+                level.addParticle(ParticleTypes.CRIT, pos.getX(), pos.getY(),pos.getZ(),1,1,1);
+                level.destroyBlock(pos, true);
+                level.updateNeighborsAt(pos, level.getBlockState(pos).getBlock());
 
-                    if (block == Blocks.OBSIDIAN) {
-                        itemstack.hurtAndBreak(40, player, (p_41300_) -> {
-                            p_41300_.broadcastBreakEvent(pContext.getHand());
+                if (block == Blocks.OBSIDIAN) {
+                    itemstack.hurtAndBreak(40, player, (p_41300_) -> {
+                        p_41300_.broadcastBreakEvent(pContext.getHand());
 
-                        });
-                    } else {
-                        itemstack.hurtAndBreak(20, player, (p_41300_) -> {
-                            p_41300_.broadcastBreakEvent(pContext.getHand());
+                    });
+                } else {
+                    itemstack.hurtAndBreak(20, player, (p_41300_) -> {
+                        p_41300_.broadcastBreakEvent(pContext.getHand());
 
-                        });
-                    }
-                    return InteractionResult.SUCCESS;
-
+                    });
                 }
+            }
+            return InteractionResult.SUCCESS;
+
+        }
 
 
         return InteractionResult.PASS;
 
+    }
+
+    @Override
+    public void inventoryTick(ItemStack pStack, Level pLevel, Entity pEntity, int pSlotId, boolean pIsSelected)
+    {
+        if(pEntity instanceof Player player && pIsSelected)
+        {
+            if(UsefullStuff.DwarfClasses.canPlayerUseItem(pStack, player, IClassData.Classes.LeafLover))
+            {
+                player.displayClientMessage(new TextComponent(ChatFormatting.RED + new TranslatableComponent("tools.you_leaf_lover").getString()), true);
+            }
+        }
+        super.inventoryTick(pStack, pLevel, pEntity, pSlotId, pIsSelected);
     }
 
     @Override
@@ -104,5 +126,15 @@ public class CompanyPickAxeItem extends PickaxeItem  {
     @Override
     public boolean isBookEnchantable(ItemStack stack, ItemStack book) {
         return false;
+    }
+
+    @Override
+    public float getDestroySpeed(ItemStack pStack, BlockState pState) {
+
+        if(UsefullStuff.DwarfClasses.canItemBeUsed(pStack, IClassData.Classes.LeafLover))
+        {
+            return 0.0f;
+        }
+        return super.getDestroySpeed(pStack, pState);
     }
 }
