+++ src/main/java/net/the_goldbeards/lootdebugs/Items/Tools/PingTool/PingItem.java
@@ -1,19 +1,18 @@
 package net.the_goldbeards.lootdebugs.Items.Tools.PingTool;
 
-import net.minecraft.util.Mth;
+import net.minecraft.ChatFormatting;
+import net.minecraft.network.chat.TextComponent;
+import net.minecraft.network.chat.TranslatableComponent;
 import net.minecraft.world.InteractionHand;
 import net.minecraft.world.InteractionResultHolder;
-import net.minecraft.world.effect.MobEffects;
+import net.minecraft.world.entity.Entity;
 import net.minecraft.world.entity.LivingEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
-import net.minecraft.world.level.ClipContext;
 import net.minecraft.world.level.Level;
-import net.minecraft.world.phys.BlockHitResult;
-import net.minecraft.world.phys.Vec3;
 import net.the_goldbeards.lootdebugs.Entities.Tools.PingEntity;
-import net.the_goldbeards.lootdebugs.Sound.ModSounds;
+import net.the_goldbeards.lootdebugs.capability.Class.IClassData;
 import net.the_goldbeards.lootdebugs.util.UsefullStuff;
 
 public class PingItem extends Item {
@@ -23,64 +22,64 @@ public class PingItem extends Item {
         super(pProperties);
     }
 
-
-
-
     @Override
-    public InteractionResultHolder<ItemStack> use(Level pLevel, Player pPlayer, InteractionHand pUsedHand) {
+    public InteractionResultHolder<ItemStack> use(Level pLevel, Player pPlayer, InteractionHand pUsedHand)
+    {
 
+        ItemStack pStack = pPlayer.getItemInHand(pUsedHand);
 
-
-        pPlayer.getCooldowns().addCooldown(this,10);
-        ItemStack pingItemStack = pPlayer.getItemInHand(pUsedHand);
-        pPlayer.playSound(ModSounds.PING_MARK.get(), 12f,1f);
-
-
-
-        if(pingItemStack.hasTag())
+        if(!UsefullStuff.DwarfClasses.canPlayerUseItem(pStack, pPlayer, IClassData.Classes.LeafLover))
         {
 
-                if(pLevel.getEntity((int) UsefullStuff.ItemNBTHelper.getFloat(pingItemStack,"lootdebugs.pingitem.pingentity.id")) instanceof PingEntity PE)
-                {
-                    PE.kill();
+            if (!pLevel.isClientSide())
+            {
+                pPlayer.getCooldowns().addCooldown(this, 10);
+                ItemStack pingItemStack = pPlayer.getItemInHand(pUsedHand);
+             //   pLevel.playSound(null, pPlayer.blockPosition(), ModSounds.MARK_SOUND.get(), SoundSource.PLAYERS, 1f, 1f);
+
+                if (pingItemStack.hasTag()) {
+
+                    if (pLevel.getEntity((int) UsefullStuff.ItemNBTHelper.getFloat(pingItemStack, "lootdebugs.pingitem.pingentity.id")) instanceof PingEntity PE) {
+                        PE.kill();
+                    } else if (pLevel.getEntity((int) UsefullStuff.ItemNBTHelper.getFloat(pingItemStack, "lootdebugs.pingitem.pingentity.id")) instanceof LivingEntity LE) {
+
+                        LE.setGlowingTag(false);
+
+                    }
                 }
 
-            else if(pLevel.getEntity((int) UsefullStuff.ItemNBTHelper.getFloat(pingItemStack,"lootdebugs.pingitem.pingentity.id")) instanceof LivingEntity LE)
-            {
+                PingEntity pingEntity = new PingEntity(pPlayer, pLevel, pingItemStack);
 
-                LE.removeEffect(MobEffects.GLOWING);
-
+                pingEntity.setOwner(pPlayer);
+                pLevel.addFreshEntity(pingEntity);
+                pingEntity.shootFromRotation(pPlayer, pPlayer.getXRot(), pPlayer.getYRot(), 0.0F, 10F, 0.0F);
+                UsefullStuff.ItemNBTHelper.putFloat(pingItemStack, "lootdebugs.pingitem.pingentity.id", (int) pingEntity.getId());//save ping id in itemstack
             }
         }
-
-        PingEntity pingEntity = new PingEntity(pPlayer,pLevel, pingItemStack);
-
-        pingEntity.setOwner(pPlayer);
-        pLevel.addFreshEntity(pingEntity);
-        pingEntity.shootFromRotation(pPlayer, pPlayer.getXRot(), pPlayer.getYRot(), 0.0F, 10F, 0.0F);
-        UsefullStuff.ItemNBTHelper.putFloat(pingItemStack,"lootdebugs.pingitem.pingentity.id", (int) pingEntity.getId());//save ping id in itemstack
         return InteractionResultHolder.pass(pPlayer.getItemInHand(pUsedHand));
     }
 
+    @Override
+    public void inventoryTick(ItemStack pStack, Level pLevel, Entity pEntity, int pSlotId, boolean pIsSelected)
+    {
+        if(pEntity instanceof Player player && pIsSelected)
+        {
+            if(UsefullStuff.DwarfClasses.canPlayerUseItem(pStack, player, IClassData.Classes.LeafLover))
+            {
+                player.displayClientMessage(new TextComponent(ChatFormatting.RED + new TranslatableComponent("tools.you_leaf_lover").getString()), true);
+            }
+        }
 
+        super.inventoryTick(pStack, pLevel, pEntity, pSlotId, pIsSelected);
+    }
+
+    @Override
+    public boolean shouldCauseReequipAnimation(ItemStack oldStack, ItemStack newStack, boolean slotChanged) {
+        return false;
+    }
 
     @Override
     public int getItemStackLimit(ItemStack stack) {
         return 1;
-    }
-
-    protected static BlockHitResult getPlayerPOVHitResult(Level pLevel, Player pPlayer, ClipContext.Fluid pFluidMode) {
-        float f = pPlayer.getXRot();
-        float f1 = pPlayer.getYRot();
-        Vec3 vec3 = pPlayer.getEyePosition();
-        float f2 = Mth.cos(-f1 * ((float)Math.PI / 180F) - (float)Math.PI);
-        float f3 = Mth.sin(-f1 * ((float)Math.PI / 180F) - (float)Math.PI);
-        float f4 = -Mth.cos(-f * ((float)Math.PI / 180F));
-        float f5 = Mth.sin(-f * ((float)Math.PI / 180F));
-        float f6 = f3 * f4;
-        float f7 = f2 * f4;
-        double d0 = 100;//Distance
-        Vec3 vec31 = vec3.add((double)f6 * d0, (double)f5 * d0, (double)f7 * d0);
-        return pLevel.clip(new ClipContext(vec3, vec31, ClipContext.Block.OUTLINE, pFluidMode, pPlayer));
     }
 }
